\name{beetles}
\alias{beetles}
\docType{data}
\title{
Beetle occurrence data from (Ulrich and Zalewski 2006).
}
\description{
Occurrence data for 71 species from 17 sites. Columns are sites, rows are species. 1's are presences and 0's are absences.}
\usage{data(beetles)}
\format{
  A data frame with 71 observations on the following 17 variables.

}

\references{
Ulrich W, Zalewski M (2006). Abundance and Co-occurrence Patterns of Core and Satellite Species of Ground Beetles on Small Lake Islands. Oikos, 114, 338-348.
}
\examples{
data(beetles)
}
\keyword{datasets}
