\name{coprob}
\alias{coprob}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function, not intended for use by the user.
}
\description{Calculate co-occurrence probabilities}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
