\name{finches}
\alias{finches}
\docType{data}
\title{
Finch occurrence data from (Sanderson 2000).
}
\description{
Occurrence data for 13 species from 17 sites. Columns are sites, rows are species. 1's are presences and 0's are absences.
}
\usage{data(finches)}
\format{
  A data frame with 13 observations on the following 17 variables.
}

\references{
Sanderson JG (2000). Testing Ecological Patterns: A Well-known Algorithm from Computer Science Aids the Evaulation of Species Distributions. American Scientist, 88, pp. 332-339.
}
\examples{
data(finches)
}
\keyword{datasets}
