% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.cooccurrence}
\alias{print.summary.cooccurrence}
\title{Print a cooccurrence summary}
\usage{
\method{print}{summary.cooccurrence}(x, n = 5L, ...)
}
\arguments{
\item{x}{A \code{summary.cooccurrence} object.}

\item{n}{Integer. Number of top nodes/groups to show. Default 5.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a cooccurrence summary
}
