\name{rodents}
\alias{rodents}
\docType{data}
\title{
Rodent occurrence data from (Brown and Kurzius 1987).}
\description{
Occurrence data for 16 species from 39 sites. Columns are sites, rows are species. 1's are presences and 0's are absences.}
\usage{data(rodents)}
\format{
A data frame with 16 observations on the following 39 variables.
}
\references{
Brown JH, Kurzius MA (1987). Composition of Desert Rodent Faunas: Combinations of Coexisting Species. Annales Zoologici Fennici, 24, 227-237.}
\examples{
data(rodents)
}
\keyword{datasets}
